Updates auto generated id logic in accordion item wc
